Added a simple paddle AI
I created an ai variable in Paddle that I use to toggle the ai on and off when the appropriate players control key is pressed. When the ai is on player movement keys are not checked and instead we run the Paddle function aimove()
aimove() simply gets the global variable ball y location and sets the paddles location to the same as the y location. Closes #1